@@ -13,7 +13,7 @@ setup(name='pytest-repeat',
       install_requires=['pytest>=3.6'],
       license='Mozilla Public License 2.0 (MPL 2.0)',
       keywords='pytest pytest repeat',
-      python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*',
+      python_requires='>=3.7',
       classifiers=[
           'Development Status :: 5 - Production/Stable',
           'Framework :: Pytest',
@@ -26,9 +26,7 @@ setup(name='pytest-repeat',
           'Topic :: Software Development :: Testing',
           'Topic :: Utilities',
           'Programming Language :: Python',
-          'Programming Language :: Python :: 2.7',
-          'Programming Language :: Python :: 3.5',
-          'Programming Language :: Python :: 3.6',
+          'Programming Language :: Python :: 3 :: Only',
           'Programming Language :: Python :: 3.7',
           'Programming Language :: Python :: 3.8',
           'Programming Language :: Python :: 3.9',
